Move pf_server_rdpgfx_init to pf_rdpgfx
