+++ server/proxy/pf_rdpgfx.h
@@ -26,6 +26,7 @@
 
 #include "pf_context.h"
 
+BOOL pf_server_rdpgfx_init(pServerContext* ps);
 void pf_rdpgfx_pipeline_init(RdpgfxClientContext* gfx, RdpgfxServerContext* server,
                              proxyData* pdata);
